requirements.txt
academictorrents==2.3.3
albumentations==2.0.8
arrow==1.3.0
Cartopy==0.24.1
clip==0.2.0
einops==0.8.1
hydra-core==1.3.2
kornia==0.8.1
matplotlib==3.10.3
natsort==8.4.0
numpy==2.3.1
omegaconf==2.3.0
pandas==2.3.1
Pillow==11.3.0
pytorch_lightning==2.5.0.post0
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.16.0
termcolor==3.1.0
timm==1.0.16
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.1
xarray==2024.11.0
